requirements.txt
cffi==1.14.4
cryptography==3.2.1
Django==2.2
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-rest-resetpassword==0.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
jwcrypto==0.8
pycparser==2.20
PyJWT==1.7.1
python-jwt==3.3.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
gunicorn==20.0.4
django-heroku==0.3.1